requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.0
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.25
Pillow==10.1.0
gunicorn==21.2.0
